backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
asyncpg==0.29.0
sqlalchemy==2.0.23

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Document Processing
pypdf==3.17.1
python-docx==1.1.0
python-pptx==0.6.23

# NLP & AI
spacy==3.7.2
transformers==4.35.2
nltk==3.8.1
sentence-transformers==2.2.2
google-generativeai==0.8.3

# Graph & Visualization
networkx==3.2.1

# Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP Client
httpx==0.25.2

# File Handling
aiofiles==24.1.0

# Telegram Bot
python-telegram-bot==20.7

# Logging & Monitoring
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Production server
gunicorn==21.2.0

# Code Quality
black==23.12.0
isort==5.13.2
flake8==6.1.0
mypy==1.7.1
